Move save-game file chooser logic to SaveGameFileChooser
