+++ game-core/src/main/java/games/strategy/triplea/ui/TripleAFrame.java
@@ -32,6 +32,7 @@ import games.strategy.engine.framework.LocalPlayers;
 import games.strategy.engine.framework.ServerGame;
 import games.strategy.engine.framework.startup.ui.InGameLobbyWatcherWrapper;
 import games.strategy.engine.framework.system.SystemProperties;
+import games.strategy.engine.framework.ui.SaveGameFileChooser;
 import games.strategy.engine.history.HistoryNode;
 import games.strategy.engine.history.Round;
 import games.strategy.engine.history.Step;
@@ -2192,7 +2193,7 @@ public final class TripleAFrame extends JFrame implements QuitHandler {
                       JOptionPane.INFORMATION_MESSAGE);
                   data.acquireReadLock();
                   try {
-                    final File f = TripleAMenuBar.getSaveGameLocation(TripleAFrame.this);
+                    final File f = SaveGameFileChooser.getSaveGameLocation(TripleAFrame.this);
                     if (f != null) {
                       try (FileOutputStream fileOutputStream = new FileOutputStream(f)) {
                         final GameData datacopy = GameDataUtils.cloneGameData(data, true);
